serialization issues/ i18n / imports
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@4230 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/connect/org.wcs.smart.connect/src/org/wcs/smart/connect/ui/server/configure/ConnectServerDialog.java
@@ -295,9 +295,9 @@ public class ConnectServerDialog extends TitleAreaDialog {
 					}else if (rec != null && currentRevision == null){
 						lblLocalChanges.setText("ERROR");
 					}else if (rec != null && currentRevision != null){
-						if (currentRevision > rec.getEndRevision()){
+						if (currentRevision.longValue() > rec.getEndRevision().longValue()){
 							lblLocalChanges.setText("Yes");
-						}else if (currentRevision == rec.getEndRevision()){
+						}else if (currentRevision.longValue() == rec.getEndRevision().longValue()){
 							lblLocalChanges.setText("No");
 						}else{
 							lblLocalChanges.setText("ERROR");
